improved player inheritance I think
big changes in character player knight
