+++ ARPG-CSE5912-LTS/Assets/Scripts/Controllers/Player/PlayerController/Combat/AttackSystem/EnemyTarget.cs
@@ -2,10 +2,9 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace ARPG.Combat
-{
+
     public class EnemyTarget : MonoBehaviour
     {
 
     }
-}
+
